Moved classes to Asset folder.
